requirements.txt
tabula-py
pandas
flask
jpype1